+++ sickrage/core/searchers/daily_searcher.py
@@ -143,9 +143,9 @@ class DailySearcher(object):
 
                 if curStatus != WANTED:
                     if bestQualities:
-                        if curQuality in bestQualities or curQuality > max(bestQualities):
+                        if curQuality in bestQualities or curQuality != Quality.UNKNOWN and curQuality > max(bestQualities):
                             continue
-                    elif curQuality in anyQualities or curQuality > max(anyQualities):
+                    elif curQuality in anyQualities or curQuality != Quality.UNKNOWN and curQuality > max(anyQualities):
                         continue
 
                 # only fetch if not archive on first match